Refactor: move functions to Agda.TypeChecking.Rules.Record.Cubical
